adicionando valores

+++ src/main/java/com/lojaVirtual/lojaVirtual/services/produto/interfaces/CategoriaService.java
@@ -15,5 +15,5 @@ public interface CategoriaService {
 
     boolean criarCategoria(CategoriaDTO categoria);
 
-    boolean deletaCategoria(int id);
+    boolean deletaCategoria(long id);
 }
